requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
python-telegram-bot>=13.0
python-dotenv>=0.19.0